feat(a4): shared grammar layer + v0 mappers (unified rule grammar, step 1)
First A4 increment per docs/design/unified-rule-grammar.md §4.1:

packages/shared/src/rules/ (the data half, §5.1):
- grammar.ts — RuleDocument v1 envelope: RulePredicate {kind,params} +
  all/any RuleGroup + recursive structural Zod schemas. Depth-1 write
  check separate from the schema (recursion is a future unlock).
  Renamed from the doc's Condition/* to avoid colliding with the live
  v0 Condition export during the strangler period
- field-match.ts — notifications' style as one kind; operator semantics
  documented verbatim incl. the String(undefined)→"undefined" edge
- contexts.ts — domain→legal-kinds registry (cleanup/auto-tag share the
  52-kind criteria vocabulary; notifications = field_match only) with
  the three-tier legality model from §2.2-as-amended
- criteria.ts — rule-criteria.ts physically moved here (subsumed, not
  paralleled); types/rule-criteria.ts is now a re-export shim — zero
  consumer churn

apps/api/src/lib/rules/v0-mappers.ts (the behavior half):
- detectDocumentVersion (v1 / v0-criteria / v0-notifications)
- mapCriteriaV0ToDocument — single + composite, near-identity; purely
  structural (retired tautulli_* kinds map fine — legality is the
  engine's tier-2/3 concern, pinned by test)
- mapNotificationsV0ToDocument — flat implicit-AND → one all-group of
  field_match; empty array → vacuous-true all-group (2.x semantics)

Tests: 12 grammar tests (incl. the vocabulary/schema-map drift guard
from §6.2) + 22 mapper fixture tests (template-derived v0 shapes).
Also carries the §6 implementation-checkpoint amendment from the
pre-implementation ultrathink (kind-legality moved out of parse time;
fixture re-sourcing; 52-kind count).
