+++ packages/shared/src/index.ts
@@ -1,1 +1,2 @@
+export * from "./rules/index";
 export * from "./types/index";
